move ui.navigation to store/ui/navigation
