+++ client/components/CountrySelect/CountrySelect.tsx
@@ -12,7 +12,7 @@ import { useUser } from '@client/store/user'
 
 import { useTranslation } from 'react-i18next'
 import Icon from '@client/components/Icon'
-import { useNavigationVisible } from '@client/store/ui'
+import { useNavigationVisible } from '@client/store/ui/navigation'
 import LinkLanding from './LinkLanding'
 import CountryList from './CountryList'
 import ToggleNavigationControl from './ToggleNavigationControl'
