+++ client/components/CountrySelect/ToggleNavigationControl/ToggleNavigationControl.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 
 import { useIsAdmin, useIsHome } from '@client/hooks'
 import Icon from '@client/components/Icon'
-import { UiActions } from '@client/store/ui'
+import { NavigationActions } from '@client/store/ui/navigation'
 import { useAppDispatch } from '@client/store'
 
 const ToggleNavigationControl: React.FC = () => {
@@ -16,7 +16,7 @@ const ToggleNavigationControl: React.FC = () => {
       className="btn app-header__toggle-navigation-visible"
       disabled={isHome || isAdmin}
       onClick={() => {
-        dispatch(UiActions.toggleNavigationVisible())
+        dispatch(NavigationActions.toggleNavigationVisible())
       }}
     >
       <Icon className="icon-sub" name="menu-left" />
