+++ client/components/Navigation/NavigationMobile/NavigationMobile.tsx
@@ -3,7 +3,7 @@ import React from 'react'
 import classNames from 'classnames'
 
 import Icon from '@webapp/components/icon'
-import { UiActions, useNavigationVisible } from '@client/store/ui'
+import { NavigationActions, useNavigationVisible } from '@client/store/ui/navigation'
 import { useAppDispatch } from '@client/store'
 import NavAssessment from '../NavAssessment'
 
@@ -15,7 +15,7 @@ const NavigationMobile: React.FC = () => {
     <div className={classNames('navM', 'no-print', { open: navigationVisible })}>
       <button
         className="btn btn-s btn-secondary navM__btnClose"
-        onClick={() => dispatch(UiActions.updateNavigationVisible(false))}
+        onClick={() => dispatch(NavigationActions.updateNavigationVisible(false))}
         type="button"
       >
         <Icon name="remove" className="icon-close icon-24" />
