+++ client/pages/Assessment/Assessment.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import Navigation from '@client/components/Navigation'
 import { Switch } from 'react-router-dom'
-import { useNavigationVisible } from '@client/store/ui'
+import { useNavigationVisible } from '@client/store/ui/navigation'
 
 const Assessment: React.FC = () => {
   const navigationVisible = useNavigationVisible()
