+++ client/store/RootState.ts
@@ -1,4 +1,4 @@
-import { UiState } from '@client/store/ui'
+import { NavigationState } from './ui/navigation/stateType'
 import { AssessmentState } from './assessment/stateType'
 import { LoginState } from './login/stateType'
 import { UserState } from './user/stateType'
@@ -7,5 +7,7 @@ export type RootState = {
   assessment: AssessmentState
   login: LoginState
   user: UserState
-  ui: UiState
+  ui: {
+    navigation: NavigationState
+  }
 }
