+++ client/store/rootReducer.ts
@@ -1,11 +1,14 @@
+import { combineReducers } from 'redux'
 import AssessmentSlice from './assessment/slice'
 import LoginSlice from './login/slice'
 import UserSlice from './user/slice'
-import UiSlice from './ui/slice'
+import NavigationSlice from './ui/navigation/slice'
 
 export default {
   assessment: AssessmentSlice,
   login: LoginSlice,
   user: UserSlice,
-  ui: UiSlice,
+  ui: combineReducers({
+    navigation: NavigationSlice,
+  }),
 }
