--- client/store/ui/index.ts
@@ -1,3 +1,0 @@
-export type { UiState } from './stateType'
-export { UiActions } from './slice'
-export { useNavigationVisible } from './hooks'
