+++ client/store/ui/navigation/index.ts
@@ -1,0 +1,3 @@
+export type { NavigationState } from './stateType'
+export { NavigationActions } from './slice'
+export { useNavigationVisible } from './hooks'
