+++ client/store/ui/navigation/stateType.ts
@@ -1,0 +1,3 @@
+export type NavigationState = {
+  visible: boolean
+}
