--- client/store/ui/stateType.ts
@@ -1,5 +1,0 @@
-export type UiState = {
-  navigation: {
-    visible: boolean
-  }
-}
